refactor(cli): split agent.rs into submodules + extract tests — no behavior change
Pure mechanical split of the ~4.3k-line stella-cli/src/agent.rs. All items
keep their names and signatures; external paths (crate::agent::X) resolve
unchanged via glob re-exports from the root. No logic or test-assertion edits.

- agent_tests.rs: the #[cfg(test)] module, moved out via #[path] so the
  forcing test does not count it against agent.rs + agent/*.rs.
- agent/prompt.rs: system-prompt assembly + file-tree rendering.
- agent/engine.rs: engine tuning + pipeline provider wiring.
- agent/tools.rs: registry options + workspace port adapters.
- agent/goal.rs: goal-driven turns.

Only widening: nine private items lifted to pub(crate) for cross-module or
test access (SYSTEM_PROMPT, STELLA_AB_RECALL_RATE, assemble_system_prompt,
build_pipeline_system_prompt, render_file_tree, provider_family,
resolve_cross_family_judge, run_raw_one_shot, run_goal_turn).
